feat(src): Implement account creation backend code
Updated alembic to allow the for the Fastapi swagger test to
successfully work. Also changed file naming to help with proper code
importing.
