+++ src/backend/alembic/versions/2881204d8f1a_add_is_active_column.py
@@ -1,0 +1,34 @@
+"""Add is_active column
+
+Revision ID: 2881204d8f1a
+Revises: 0168026960a3
+Create Date: 2024-03-05 21:09:39.409638
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '2881204d8f1a'
+down_revision: Union[str, None] = '0168026960a3'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('users', sa.Column('is_active', sa.Boolean(), nullable=True))
+    op.drop_index('ix_users_name', table_name='users')
+    op.drop_column('users', 'name')
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('users', sa.Column('name', sa.VARCHAR(), nullable=True))
+    op.create_index('ix_users_name', 'users', ['name'], unique=False)
+    op.drop_column('users', 'is_active')
+    # ### end Alembic commands ###
