+++ src/backend/app/database.py
@@ -3,9 +3,11 @@ from sqlalchemy.orm import sessionmaker
 from sqlalchemy.ext.declarative import declarative_base
 
 # setting up the database
-engine = create_engine('sqlite:///./bank.db')
+database = 'sqlite:///./bank.db'
+engine = create_engine(
+    database, connect_args={"check_same_thread": False}
+)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 Base = declarative_base()
 
-Base.metadata.create_all(bind=engine)
