+++ src/backend/app/models.py
@@ -1,9 +1,8 @@
 from sqlalchemy import Boolean, Column, Integer, String, Date
-#from sqlalchemy import relationship
+#from sqlalchemy.orm import relationship
 
-from typing import Optional
 
-from .database import Base
+from app.database import Base
 
 
 
@@ -12,12 +11,9 @@ class User(Base):
     __tablename__ = 'users'
 
     id = Column(Integer, primary_key=True, index=True)
-    #name = Column(String, index=True)
     email = Column(String, unique=True, index=True)
     password = Column(String)
     is_active = Column(Boolean, default=True)
-    #postCode = Column(String)
-    # DoB = Column(Date)
