+++ src/backend/app/schemas.py
@@ -3,9 +3,7 @@ from pydantic import BaseModel
 
 class UserBase(BaseModel):
     email: str
-    #name: str 
-    #postCode: str
-    #DoB: Optional[Date] = None
+
 
 class UserCreate(UserBase):
     password: str
@@ -13,9 +11,6 @@ class UserCreate(UserBase):
 class User(UserBase):
     id: int
     is_active: bool
-    #name: str
-    #postCode: str
-    #DoB: Optional[Date] = None
 
     class Config:
         from_attributes = True
