+++ src/backend/main.py
@@ -2,8 +2,8 @@ from fastapi import Depends, FastAPI, HTTPException
 from sqlalchemy.orm import Session
 
 
-from code import crud, models, schemas
-from code.database import SessionLocal, engine
+from app import crud, models, schemas
+from app.database import SessionLocal, engine
 
 models.Base.metadata.create_all(bind=engine)
 
@@ -30,7 +30,7 @@ def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
 
 
 @app.get("/users/", response_model=list[schemas.User])
-def read_users(skip:int=0, limit: int=100, db: Session = Depends(get_db)):
+def read_users(skip: int=0, limit: int=100, db: Session = Depends(get_db)):
     users = crud.get_users(db, skip=skip, limit=limit)
     return users #could be this?
